Improve inter-thread communication
- Add type/name pair to thread messages
- Decompose message handlers
