+++ lib/application.js
@@ -7,6 +7,7 @@ const fs = require('fs');
 const { MessageChannel, parentPort, threadId } = require('worker_threads');
 const metavm = require('metavm');
 const metawatch = require('metawatch');
+const metautil = require('metautil');
 const { Interfaces } = require('./interfaces.js');
 const { Modules } = require('./modules.js');
 const { Resources } = require('./resources.js');
@@ -196,7 +197,8 @@ class Application extends events.EventEmitter {
 
   async invoke({ method, args }) {
     const { port1, port2 } = new MessageChannel();
-    const msg = { type: 'invoke', data: { method, args }, port: port1 };
+    const data = { method, args };
+    const msg = { type: 'invoke', name: 'request', data, port: port1 };
     return new Promise((resolve, reject) => {
       port2.on('message', ({ error, data }) => {
         if (error) reject(error);
@@ -207,4 +209,27 @@ class Application extends events.EventEmitter {
   }
 }
 
-module.exports = new Application();
+const application = new Application();
+
+if (parentPort) {
+  parentPort.on('message', async ({ type, name, data, port }) => {
+    if (type !== 'invoke' || name !== 'request') return;
+    const { method, args } = data;
+    const handler = metautil.namespaceByPath(application.sandbox, method);
+    if (!handler) {
+      port.postMessage({ error: new Error('Handler not found') });
+      return;
+    }
+    try {
+      const result = await handler(args);
+      port.postMessage({ data: result });
+    } catch (err) {
+      port.postMessage({ error: err });
+      application.console.error(err.stack);
+    } finally {
+      parentPort.postMessage({ type: 'invoke', name: 'done' });
+    }
+  });
+}
+
+module.exports = application;
