+++ lib/worker.js
@@ -7,7 +7,6 @@ const { Config } = require('metaconfiguration');
 const { Logger } = require('metalog');
 const { Server } = require('metacom');
 const metavm = require('metavm');
-const metautil = require('metautil');
 const { notLoaded } = require('./dependencies.js');
 const application = require('./application.js');
 
@@ -76,30 +75,11 @@ const application = require('./application.js');
   console.info(`Application started in worker ${threadId}`);
   parentPort.postMessage({ type: 'event', name: 'started' });
 
-  parentPort.on('message', async ({ type, name, data, port }) => {
-    if (type === 'event') {
-      if (name === 'stop') {
-        if (application.finalization) return;
-        console.info(`Graceful shutdown in worker ${threadId}`);
-        await application.shutdown();
-        process.exit(0);
-      }
-      return;
-    }
-    if (type === 'invoke') {
-      const { method, args } = data;
-      const handler = metautil.namespaceByPath(application.sandbox, method);
-      if (!handler) {
-        port.postMessage({ error: new Error('Handler not found') });
-        return;
-      }
-      try {
-        const result = await handler(args);
-        port.postMessage({ data: result });
-      } catch (err) {
-        port.postMessage({ error: err });
-        application.console.error(err.stack);
-      }
-    }
+  parentPort.on('message', async ({ type, name }) => {
+    if (type !== 'event' || name !== 'stop') return;
+    if (application.finalization) return;
+    console.info(`Graceful shutdown in worker ${threadId}`);
+    await application.shutdown();
+    process.exit(0);
   });
 })();
